server/requirements.txt
Flask==2.0.1
Flask-SocketIO==5.1.1
python-dotenv==0.19.1
google-generativeai==0.3.0
torch==2.0.0
python-weather==1.0.1
googlemaps==4.6.0
websockets==10.3
googlesearch-python==1.1.0
aiohttp==3.8.1
beautifulsoup4==4.10.0
lxml==4.6.3
requests==2.26.0
eventlet==0.33.0
RealtimeSTT==0.1.5
RealtimeTTS==0.2.0
pyaudio==0.2.11
opencv-python==4.5.3.56
pillow==8.3.2
mss==6.1.0
psutil==5.8.0
GPUtil==1.4.0
elevenlabs==0.2.10
